Move constants from plugin to core
